fix(jvm): make BPF map headers clang-tidy safe
Split shared USDT and JVM BPF structs/constants into type-only headers so clang-tidy can parse the new map headers as standalone translation units.

Include vmlinux.h plus the relevant type headers from the split map definitions, which provides BPF integer typedefs and key/value types without recursive includes.

This fixes the Ubuntu clang-tidy-19 failure where obi_usdt_* and JVM sample map headers cascaded into bpf_helper_defs.h unknown __u32/__u64 diagnostics.

Validation:

- clang-tidy focused checks for the new standalone map/type headers

- clang-tidy focused misc-unused-parameters check for bpf/generictracer/jvm.c

- clang -target bpf compile of bpf/generictracer/generictracer.c for x86 and arm64

- make vanity-import-check

- go test ./pkg/ebpf/common ./pkg/appolly/app/runtime ./pkg/ebpf ./pkg/internal/ebpf/generictracer ./pkg/internal/ebpf/gotracer

- go tool -modfile=internal/tools/go.mod golangci-lint run ./pkg/ebpf --timeout=6m

- make compile
