+++ bpf/common/usdt.h
@@ -7,52 +7,12 @@
 #include <bpfcore/bpf_helpers.h>
 #include <bpfcore/bpf_tracing.h>
 #include <common/pin_internal.h>
+#include <common/usdt_types.h>
 #include <pid/pid.h>
 
 #ifndef barrier_var
 #define barrier_var(var) asm volatile("" : "+r"(var))
 #endif
-
-enum { k_obi_usdt_max_args = 12, k_obi_usdt_max_spec_cnt = 256, k_obi_usdt_max_ip_cnt = 1024 };
-
-enum obi_usdt_arg_type {
-    k_obi_usdt_arg_const = 0,
-    k_obi_usdt_arg_reg = 1,
-    k_obi_usdt_arg_reg_deref = 2,
-    k_obi_usdt_arg_sib = 3,
-};
-
-enum obi_usdt_arg_error {
-    k_obi_usdt_arg_err_no_spec = -2,
-    k_obi_usdt_arg_err_out_of_range = -3,
-    k_obi_usdt_arg_err_bad_type = -4,
-    k_obi_usdt_arg_err_bad_size = -5,
-    k_obi_usdt_arg_err_bad_reg = -6,
-    k_obi_usdt_arg_err_bad_scale = -7,
-};
-
-struct obi_usdt_arg_spec {
-    u64 val_off;
-    s16 reg_off;
-    s16 idx_reg_off;
-    u8 arg_type;
-    u8 scale_bitshift;
-    u8 arg_signed;
-    u8 arg_bitshift;
-};
-
-struct obi_usdt_spec {
-    struct obi_usdt_arg_spec args[k_obi_usdt_max_args];
-    u64 cookie;
-    u16 arg_cnt;
-    u8 _pad[6];
-};
-
-struct obi_usdt_ip_key {
-    u32 pid;
-    u32 ns;
-    u64 ip;
-};
 
 #include <common/maps/obi_usdt_ip_to_spec_id.h>
 #include <common/maps/obi_usdt_specs.h>
