+++ bpf/generictracer/maps/jvm_mem_pool_samples.h
@@ -3,8 +3,10 @@
 
 #pragma once
 
+#include <bpfcore/vmlinux.h>
 #include <bpfcore/bpf_helpers.h>
 #include <common/pin_internal.h>
+#include <generictracer/types/jvm.h>
 
 struct {
     __uint(type, BPF_MAP_TYPE_LRU_HASH);
